test: move token helper into hono integration test folder
