+++ integration-test/mockOauth2Client.ts
@@ -4,7 +4,7 @@ import {
   IOAuth2Client,
 } from "../src/services/oauth2-client";
 import { createToken } from "../src/utils/jwt";
-import { getCertificate } from "./helpers/token";
+import { getCertificate } from "../test/integration/helpers/token";
 
 function createTokenExample(payload: {
   [key: string]: string | string[] | number | boolean;
