+++ scripts/create-token.ts
@@ -1,3 +1,3 @@
-import { getAdminToken } from "../integration-test/helpers/token";
+import { getAdminToken } from "../test/integration/helpers/token";
 
 getAdminToken().then((token) => console.log(`Token: ${token}`));
