+++ test/fixtures/context.ts
@@ -15,7 +15,7 @@ import { mockedR2Bucket } from "./mocked-r2-bucket";
 import { EmailOptions } from "../../src/services/email/EmailOptions";
 import { Var } from "../../src/types/Var";
 import createAdapters from "../../src/adapters/in-memory";
-import { getCertificate } from "../../integration-test/helpers/token";
+import { getCertificate } from "../integration/helpers/token";
 import { sendLink, sendCode } from "../../src/controllers/email";
 import { Ticket } from "../../src/types/Ticket";
 import { OTP } from "../../src/types/OTP";
