+++ test/integration/keys.spec.ts
@@ -1,6 +1,6 @@
 import { testClient } from "hono/testing";
 import { tsoaApp } from "../../src/app";
-import { getAdminToken } from "../../integration-test/helpers/token";
+import { getAdminToken } from "./helpers/token";
 import { getEnv } from "./helpers/test-client";
 
 describe("keys", () => {
