+++ test/integration/management-api/users-by-email.spec.ts
@@ -1,6 +1,6 @@
 import { testClient } from "hono/testing";
 import { tsoaApp } from "../../../src/app";
-import { getAdminToken } from "../../../integration-test/helpers/token";
+import { getAdminToken } from "../helpers/token";
 import { getEnv } from "../helpers/test-client";
 import { UserResponse } from "../../../src/types/auth0";
 import { Identity } from "../../../src/types/auth0/Identity";
